refactor: renamed studioSlice to studioReducer
